AdvancedDisplay: Refactor
* Integrate with Settings
* Remove back button
