+++ AdvancedDisplay/Android.mk
@@ -9,7 +9,29 @@ LOCAL_PACKAGE_NAME := AdvancedDisplay
 LOCAL_CERTIFICATE := platform
 
 LOCAL_STATIC_JAVA_LIBRARIES := \
+    android-support-v14-preference \
+    android-support-v7-appcompat \
+    android-support-v7-preference \
+    android-support-v7-recyclerview \
     org.cyanogenmod.platform.internal
+
+LOCAL_PROGUARD_FLAG_FILES := proguard.flags
+
+LOCAL_RESOURCE_DIR := \
+    $(LOCAL_PATH)/res \
+    frameworks/support/v14/preference/res \
+    frameworks/support/v7/appcompat/res \
+    frameworks/support/v7/preference/res \
+    frameworks/support/v7/recyclerview/res
+
+LOCAL_AAPT_FLAGS := \
+    --auto-add-overlay \
+    --extra-packages android.support.v14.preference \
+    --extra-packages android.support.v7.appcompat \
+    --extra-packages android.support.v7.preference \
+    --extra-packages android.support.v7.recyclerview
+
+include frameworks/base/packages/SettingsLib/common.mk
 
 include $(BUILD_PACKAGE)
