+++ AdvancedDisplay/proguard.flags
@@ -1,0 +1,5 @@
+-keepclasseswithmembers class * {
+    public <init>(android.content.Context, android.util.AttributeSet);
+}
+
+-keep class ** extends android.support.v14.preference.PreferenceFragment
